+++ keyboard_rewrite/src/lib/y2keyboard/dialogs/layout_selector.rb
