+++ keyboard_rewrite/src/lib/y2keyboard/keyboard_layout.rb
