+++ keyboard_rewrite/test/keyboard_layout_spec.rb
@@ -1,5 +1,5 @@
 require_relative "test_helper"
-require "y2_keyboard/keyboard_layout"
+require "y2keyboard/keyboard_layout"
 
 describe Y2Keyboard::KeyboardLayout do
   describe ".all" do
